backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
motor==3.7.1
pymongo==4.16.0
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
